Setup as legit python package Fixes #2
