+++ src/incremental_thinking/utils.py
@@ -1,0 +1,9 @@
+import random
+
+
+def prob_generator(number):
+    return 1 / (number + 0.2)
+
+
+def convert_prob_to_bool(prob):
+    return random.random() < prob
